+++ products/models.py
@@ -22,15 +22,3 @@ class Product(models.Model):
     def __str__(self) -> str:
         return self.Product_Name
     
-#User
-class User(models.Model):
-    name = models.CharField(max_length=20, null=False)
-    cc = models.IntegerField()
-    username = models.CharField(max_length=20, null=False)
-    email = models.CharField(max_length=30, null=False)
-    password = models.CharField(max_length=30, null=False)
-    type = models.CharField(max_length=15)
-    def __str__(self) -> str:
-        return self.name
-
-
